requirements.txt
asyncpg
git+https://github.com/Rapptz/discord.py@master#egg=discord.py[voice]
git+https://github.com/Gorialis/jishaku@master#egg=jishaku
git+https://github.com/Zenrac/TextToOwO
psutil
git+https://github.com/Rapptz/discord-ext-menus
Quart
requests_oauthlib
dateparser
pendulum
pyjwt
lavalink
dblpy